requirements.txt
streamlit
langchain-openai
pinecone-client
pandas
python-dotenv
requests

